Cap network timeouts at 32-bit integers, because some operating systems (*cough* Mac OS X *cough*) have undefined maximum timeout values for select.  Keeping everything 32-bit means we shouldn't hit this undefined maximum, but it does mean the server might have to wake up every 49 days.  ("A wait of 49 days should be sufficient to keep it from busywaiting" -- Jeffrey)
git-svn-id: https://legesmotus.svn.sourceforge.net/svnroot/legesmotus/legesmotus/trunk@376 a5a5d5d7-22d1-44d1-b37a-67d3d99fcfd7

+++ common/UDPSocket.hpp
@@ -28,7 +28,7 @@ public:
 
 	// return true as soon as a packet is ready for reading
 	// will wait for up to wait_time milliseconds for a packet, after which it will return false
-	bool	has_packets(uint64_t wait_time =0); 
+	bool	has_packets(uint32_t wait_time =0); 
 
 	bool	send(const UDPPacket&);
 	bool	recv(UDPPacket&);
